Code refactoring, mostly grid containers
