+++ src/ClassicUO.Client/Game/UI/Gumps/GridContainerListView.cs
@@ -5,11 +5,11 @@ namespace ClassicUO.Game.UI.Gumps
 {
     public partial class GridContainer
     {
-        private const int LIST_ROW_HEIGHT = 40;
-        private const int LIST_ICON_SIZE = 40;
-        private const int LIST_NAME_MAX_CHARS = 20;
-        private const int LIST_COLUMN_WIDTH = 200;
-        private const int LIST_COLUMN_GAP = 8;
+        public const int LIST_ICON_SIZE = 40;
+        public const int LIST_NAME_MAX_CHARS = 20;
+        public const int LIST_ROW_HEIGHT = 40;
+        public const int LIST_COLUMN_WIDTH = 200;
+        public const int LIST_COLUMN_GAP = 8;
 
         private enum GridContainerViewMode
         {
@@ -41,7 +41,7 @@ namespace ClassicUO.Game.UI.Gumps
             }
         }
 
-        private bool IsListView => EffectiveViewMode == GridContainerViewMode.List;
+        public bool IsListView => EffectiveViewMode == GridContainerViewMode.List;
 
         private void InitializeListView() => EventSink.OPLOnReceive += OnListViewOplReceived;
