+++ src/ClassicUO.Client/Game/UI/Gumps/GridContainers/Enums/BorderStyle.cs
@@ -1,0 +1,13 @@
+namespace ClassicUO.Game.UI.Gumps;
+
+public enum BorderStyle
+{
+    Default,
+    Style1,
+    Style2,
+    Style3,
+    Style4,
+    Style5,
+    Style6,
+    Style7
+}
