+++ src/ClassicUO.Client/Game/UI/Gumps/GridContainers/Enums/GridSortMode.cs
@@ -1,0 +1,7 @@
+namespace ClassicUO.Game.UI.Gumps;
+
+public enum GridSortMode
+{
+    GraphicAndHue = 0,
+    Name = 1
+}
