+++ src/ClassicUO.Client/Game/UI/Gumps/GridContainers/Enums/LowContrastHighlightStyle.cs
@@ -1,0 +1,7 @@
+namespace ClassicUO.Game.UI.Gumps;
+
+public enum LowContrastHighlightStyle
+{
+    Border,
+    Spotlight
+}
